requirements.txt
discord.py
loguru
python-dotenv
yt-dlp
pydub
ffmpeg-python
PyNaCl
pydantic